requirements.txt
# Google AI
google-genai>=0.3.0
google-adk>=0.5.0
google-cloud-aiplatform>=1.40.0

# Web Framework
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pathlib>=1.0.1

# Document processing
python-docx>=0.8.11
openpyxl>=3.10.0
pypdf>=3.17.0

# Async
aiofiles>=23.1.0
